add mongo autoconfig
