+++ ezyfox-boot-autoconfigure/src/test/java/com/tvd12/ezyfox/boot/test/mongodb/MongoConfigurationTest.java
