+++ ezyfox-boot-autoconfigure/src/test/resources/AllTests.tng.xml
