+++ pom.xml
@@ -49,6 +49,6 @@
 		</developer>
 	</developers>
 	<modules>
-		<module>ezyfox-boot-autorconfigure</module>
+		<module>ezyfox-boot-autoconfigure</module>
 	</modules>
 </project>
